Use post ID instead of URL to load editor
Then we can use the post ID to query the slug and content from the API. We can
then use that data to apply changes from the edited markup to the post.

This changes the editableSelector to `.warpedit-editable`, and the selector must
be used inside the post's content (rather than only in the rendered page).